Split network samples into multiple files to make them easier to understand.
