+++ network/loadbalancer.go
@@ -1,0 +1,143 @@
+// Copyright (c) Microsoft and contributors.  All rights reserved.
+//
+// This source code is licensed under the MIT license found in the
+// LICENSE file in the root directory of this source tree.
+
+package network
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/Azure-Samples/azure-sdk-for-go-samples/helpers"
+	"github.com/Azure-Samples/azure-sdk-for-go-samples/iam"
+	"github.com/Azure/azure-sdk-for-go/services/network/mgmt/2017-09-01/network"
+	"github.com/Azure/go-autorest/autorest"
+	"github.com/Azure/go-autorest/autorest/to"
+)
+
+// Load balancers
+
+func getLBClient() network.LoadBalancersClient {
+	token, _ := iam.GetResourceManagementToken(iam.AuthGrantType())
+	lbClient := network.NewLoadBalancersClient(helpers.SubscriptionID())
+	lbClient.Authorizer = autorest.NewBearerAuthorizer(token)
+	lbClient.AddToUserAgent(helpers.UserAgent())
+	return lbClient
+}
+
+// GetLoadBalancer gets info on a loadbalancer
+func GetLoadBalancer(ctx context.Context, lbName string) (network.LoadBalancer, error) {
+	lbClient := getLBClient()
+	return lbClient.Get(ctx, helpers.ResourceGroupName(), lbName, "")
+}
+
+// CreateLoadBalancer creates a load balancer with 2 inbound NAT rules.
+func CreateLoadBalancer(ctx context.Context, lbName, pipName string) (lb network.LoadBalancer, err error) {
+	probeName := "probe"
+	frontEndIPConfigName := "fip"
+	backEndAddressPoolName := "backEndPool"
+	idPrefix := fmt.Sprintf("/subscriptions/%s/resourceGroups/%s/providers/Microsoft.Network/loadBalancers", helpers.SubscriptionID(), helpers.ResourceGroupName())
+
+	pip, err := GetPublicIP(ctx, pipName)
+	if err != nil {
+		return
+	}
+
+	lbClient := getLBClient()
+	future, err := lbClient.CreateOrUpdate(ctx,
+		helpers.ResourceGroupName(),
+		lbName,
+		network.LoadBalancer{
+			Location: to.StringPtr(helpers.Location()),
+			LoadBalancerPropertiesFormat: &network.LoadBalancerPropertiesFormat{
+				FrontendIPConfigurations: &[]network.FrontendIPConfiguration{
+					{
+						Name: &frontEndIPConfigName,
+						FrontendIPConfigurationPropertiesFormat: &network.FrontendIPConfigurationPropertiesFormat{
+							PrivateIPAllocationMethod: network.Dynamic,
+							PublicIPAddress:           &pip,
+						},
+					},
+				},
+				BackendAddressPools: &[]network.BackendAddressPool{
+					{
+						Name: &backEndAddressPoolName,
+					},
+				},
+				Probes: &[]network.Probe{
+					{
+						Name: &probeName,
+						ProbePropertiesFormat: &network.ProbePropertiesFormat{
+							Protocol:          network.ProbeProtocolHTTP,
+							Port:              to.Int32Ptr(80),
+							IntervalInSeconds: to.Int32Ptr(15),
+							NumberOfProbes:    to.Int32Ptr(4),
+							RequestPath:       to.StringPtr("healthprobe.aspx"),
+						},
+					},
+				},
+				LoadBalancingRules: &[]network.LoadBalancingRule{
+					{
+						Name: to.StringPtr("lbRule"),
+						LoadBalancingRulePropertiesFormat: &network.LoadBalancingRulePropertiesFormat{
+							Protocol:             network.TransportProtocolTCP,
+							FrontendPort:         to.Int32Ptr(80),
+							BackendPort:          to.Int32Ptr(80),
+							IdleTimeoutInMinutes: to.Int32Ptr(4),
+							EnableFloatingIP:     to.BoolPtr(false),
+							LoadDistribution:     network.Default,
+							FrontendIPConfiguration: &network.SubResource{
+								ID: to.StringPtr(fmt.Sprintf("/%s/%s/frontendIPConfigurations/%s", idPrefix, lbName, frontEndIPConfigName)),
+							},
+							BackendAddressPool: &network.SubResource{
+								ID: to.StringPtr(fmt.Sprintf("/%s/%s/backendAddressPools/%s", idPrefix, lbName, backEndAddressPoolName)),
+							},
+							Probe: &network.SubResource{
+								ID: to.StringPtr(fmt.Sprintf("/%s/%s/probes/%s", idPrefix, lbName, probeName)),
+							},
+						},
+					},
+				},
+				InboundNatRules: &[]network.InboundNatRule{
+					{
+						Name: to.StringPtr("natRule1"),
+						InboundNatRulePropertiesFormat: &network.InboundNatRulePropertiesFormat{
+							Protocol:             network.TransportProtocolTCP,
+							FrontendPort:         to.Int32Ptr(21),
+							BackendPort:          to.Int32Ptr(22),
+							EnableFloatingIP:     to.BoolPtr(false),
+							IdleTimeoutInMinutes: to.Int32Ptr(4),
+							FrontendIPConfiguration: &network.SubResource{
+								ID: to.StringPtr(fmt.Sprintf("/%s/%s/frontendIPConfigurations/%s", idPrefix, lbName, frontEndIPConfigName)),
+							},
+						},
+					},
+					{
+						Name: to.StringPtr("natRule2"),
+						InboundNatRulePropertiesFormat: &network.InboundNatRulePropertiesFormat{
+							Protocol:             network.TransportProtocolTCP,
+							FrontendPort:         to.Int32Ptr(23),
+							BackendPort:          to.Int32Ptr(22),
+							EnableFloatingIP:     to.BoolPtr(false),
+							IdleTimeoutInMinutes: to.Int32Ptr(4),
+							FrontendIPConfiguration: &network.SubResource{
+								ID: to.StringPtr(fmt.Sprintf("/%s/%s/frontendIPConfigurations/%s", idPrefix, lbName, frontEndIPConfigName)),
+							},
+						},
+					},
+				},
+			},
+		})
+
+	if err != nil {
+		return lb, fmt.Errorf("cannot create load balancer: %v", err)
+	}
+
+	err = future.WaitForCompletion(ctx, lbClient.Client)
+	if err != nil {
+		return lb, fmt.Errorf("cannot get load balancer create or update future response: %v", err)
+	}
+
+	return future.Result(lbClient)
+}
