+++ network/nic.go
@@ -1,0 +1,144 @@
+// Copyright (c) Microsoft and contributors.  All rights reserved.
+//
+// This source code is licensed under the MIT license found in the
+// LICENSE file in the root directory of this source tree.
+
+package network
+
+import (
+	"context"
+	"fmt"
+	"log"
+
+	"github.com/Azure-Samples/azure-sdk-for-go-samples/helpers"
+	"github.com/Azure-Samples/azure-sdk-for-go-samples/iam"
+	"github.com/Azure/azure-sdk-for-go/services/network/mgmt/2017-09-01/network"
+	"github.com/Azure/go-autorest/autorest"
+	"github.com/Azure/go-autorest/autorest/to"
+)
+
+// Network Interfaces (NIC's)
+
+func getNicClient() network.InterfacesClient {
+	token, _ := iam.GetResourceManagementToken(iam.AuthGrantType())
+	nicClient := network.NewInterfacesClient(helpers.SubscriptionID())
+	nicClient.Authorizer = autorest.NewBearerAuthorizer(token)
+	nicClient.AddToUserAgent(helpers.UserAgent())
+	return nicClient
+}
+
+// CreateNIC creates a new network interface. The Network Security Group is not a required parameter
+func CreateNIC(ctx context.Context, vnetName, subnetName, nsgName, ipName, nicName string) (nic network.Interface, err error) {
+	subnet, err := GetVirtualNetworkSubnet(ctx, vnetName, subnetName)
+	if err != nil {
+		log.Fatalf("failed to get subnet: %v", err)
+	}
+
+	ip, err := GetPublicIP(ctx, ipName)
+	if err != nil {
+		log.Fatalf("failed to get ip address: %v", err)
+	}
+
+	nicParams := network.Interface{
+		Name:     to.StringPtr(nicName),
+		Location: to.StringPtr(helpers.Location()),
+		InterfacePropertiesFormat: &network.InterfacePropertiesFormat{
+			IPConfigurations: &[]network.InterfaceIPConfiguration{
+				{
+					Name: to.StringPtr("ipConfig1"),
+					InterfaceIPConfigurationPropertiesFormat: &network.InterfaceIPConfigurationPropertiesFormat{
+						Subnet: &subnet,
+						PrivateIPAllocationMethod: network.Dynamic,
+						PublicIPAddress:           &ip,
+					},
+				},
+			},
+		},
+	}
+
+	if nsgName != "" {
+		nsg, err := GetNetworkSecurityGroup(ctx, nsgName)
+		if err != nil {
+			log.Fatalf("failed to get nsg: %v", err)
+		}
+		nicParams.NetworkSecurityGroup = &nsg
+	}
+
+	nicClient := getNicClient()
+	future, err := nicClient.CreateOrUpdate(ctx, helpers.ResourceGroupName(), nicName, nicParams)
+	if err != nil {
+		return nic, fmt.Errorf("cannot create nic: %v", err)
+	}
+
+	err = future.WaitForCompletion(ctx, nicClient.Client)
+	if err != nil {
+		return nic, fmt.Errorf("cannot get nic create or update future response: %v", err)
+	}
+
+	return future.Result(nicClient)
+}
+
+// CreateNICWithLoadBalancer creats a network interface, wich is set up with a loadbalancer's NAT rule
+func CreateNICWithLoadBalancer(ctx context.Context, lbName, vnetName, subnetName, nicName string, natRule int) (nic network.Interface, err error) {
+	subnet, err := GetVirtualNetworkSubnet(ctx, vnetName, subnetName)
+	if err != nil {
+		return
+	}
+
+	lb, err := GetLoadBalancer(ctx, lbName)
+	if err != nil {
+		return
+	}
+
+	nicClient := getNicClient()
+	future, err := nicClient.CreateOrUpdate(ctx,
+		helpers.ResourceGroupName(),
+		nicName,
+		network.Interface{
+			Location: to.StringPtr(helpers.Location()),
+			InterfacePropertiesFormat: &network.InterfacePropertiesFormat{
+				IPConfigurations: &[]network.InterfaceIPConfiguration{
+					{
+						Name: to.StringPtr("pipConfig"),
+						InterfaceIPConfigurationPropertiesFormat: &network.InterfaceIPConfigurationPropertiesFormat{
+							Subnet: &network.Subnet{
+								ID: subnet.ID,
+							},
+							LoadBalancerBackendAddressPools: &[]network.BackendAddressPool{
+								{
+									ID: (*lb.BackendAddressPools)[0].ID,
+								},
+							},
+							LoadBalancerInboundNatRules: &[]network.InboundNatRule{
+								{
+									ID: (*lb.InboundNatRules)[natRule].ID,
+								},
+							},
+						},
+					},
+				},
+			},
+		})
+	if err != nil {
+		return nic, fmt.Errorf("cannot create nic: %v", err)
+	}
+
+	err = future.WaitForCompletion(ctx, nicClient.Client)
+	if err != nil {
+		return nic, fmt.Errorf("cannot get nic create or update future response: %v", err)
+	}
+
+	return future.Result(nicClient)
+}
+
+// GetNic returns an existing network interface
+func GetNic(ctx context.Context, nicName string) (network.Interface, error) {
+	nicClient := getNicClient()
+	return nicClient.Get(ctx, helpers.ResourceGroupName(), nicName, "")
+}
+
+// DeleteNic deletes an existing network interface
+func DeleteNic(ctx context.Context, nic string) (result network.InterfacesDeleteFuture, err error) {
+	nicClient := getNicClient()
+	return nicClient.Delete(ctx, helpers.ResourceGroupName(), nic)
+}
